Add get_availability to channel and listing
Not all modules (like webshop) are channel listing compatible yet.
This module introduces the ability to get the availability with quantity (optional)
or as a bucket ('in_stock', 'limited', 'out_of_stock').

The module's default implementation returns buckets and based on the
real inventory of the item available in the channel's warehouse. Downstream
modules can change this behavior